requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
Faker==22.5.1
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
img2pdf==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.4
orjson==3.9.12
packaging==23.2
pdf2image==1.17.0
pikepdf==8.11.2
pillow==10.2.0
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.1
qrcode-artistic==3.0.2
requests==2.31.0
segno==1.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.26.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
